// src/app/models/doi/doiModel.ts
export class ListValue {
    value: string;
    viewValue: string;
}